Prevent Bot Account Usage
We may have users who are using a bot to interact with Warden to automate enforcement on their FiveM server, and I want to prevent that as much as we can for Global Blacklist reasons.

If you feel this is targeting you, reach out to me in the Warden discord server so we can discuss it.